refactor: finalize exports
