+++ src/errors/PrismicRichTextError.ts
