+++ src/errors/PrismicRichTextSerializerError.ts
