+++ src/richtext/index.ts
@@ -4,19 +4,12 @@ export { asText } from "./asText";
 export { serialize } from "./serialize";
 export { wrapMapSerializer } from "./wrapMapSerializer";
 export { composeSerializers } from "./composeSerializers";
-export { filterRichTextField } from "./filterRichTextField";
+export { filterRichTextField } from "../lib/filterRichTextField";
 
 export { RichTextNodeType as Element } from "../types/value/richText";
-
-export { unstable_htmlAsRichText } from "./unstable_htmlAsRichText";
-export { unstable_markdownAsRichText } from "./unstable_markdownAsRichText";
-
-export { PrismicRichTextError } from "./errors/PrismicRichTextError";
-export { PrismicRichTextSerializerError } from "./errors/PrismicRichTextSerializerError";
 
 export type {
 	RichTextFunctionSerializer,
 	RichTextMapSerializer,
 	RichTextMapSerializerFunction,
-	RichTextHTMLMapSerializer,
 } from "./types";
