+++ test/helpers-unstable_markdownAsRichText.test.ts
@@ -1,6 +1,6 @@
 import { describe } from "vitest";
 
-import { testMarkdownAsRichTextHelper } from "../__testutils__/testAsRichTextHelper";
+import { testMarkdownAsRichTextHelper } from "./__testutils__/testAsRichTextHelper";
 
 describe("transforms markdown to rich text", () => {
 	describe("basic", () => {
